Deny #! in places where comments are not allowed
The previous behavior, of not skipping over it when skipping over
non-code tokens, meant allowing it in some places where it doesn't make
sense, even though it's not really harmful.

This change bans it in some places, though still not all of them: you
can still add a #! in places where comments are allowed. Since that's
not really harmful, I'm keeping that for now.