Hook up the smith fuzzer to Mode::PatchIdempotent
It does find most cases very quickly, but apparently it is hard for it
to create sensible patch paths. It needs to be quite lucky to generate
them with ExprField, and then duplicate the right field names that are
also used in the document. Perhaps there is a better way to explore
faster, but at least we have one entry point.

It hasn't discovered any idempotency violations yet.