Smith: Add "dup" instruction
If the same expression is needed in two parts of the program, before
this change, they had to be pushed separately, with the same sequence of
instructions. Now we can use a 'Dup' instruction instead to duplicate an
expression from within the stack. I hope that this will enable the
fuzzer to explore a bit faster.

Some context, I modified the patch implementation to be not idempotent
(or so I thought), and the fuzzer did not find a counterexample yet
after several minutes of fuzzing. But as I'm writing this, I realize
that I made a mistake there, and even my sabotaged patch command is
still idempotent. Still, this is a small modification of the smith that
can hopefully make it more efficient at exploring the search space, so
let's keep it.