Hook up the source fuzzer to Mode::PatchIdempotent
I also (temporarily) seeded the corpus with some examples of patch mode
inputs. It started exploring, but it still has not discovered a matching
path at a depth greater than 1 after about 20 minutes of fuzzing (on a
single core). I suspect this is going to take a while, but I suppose it
would get there eventually. It is at least steadily discovering new fuzz
inputs.