Support arbitrary commands in goldens, test patch
For patching, we need to have more control over what to pass on the
command line. Previously the directory implicitly controlled this, but
the time has come to make this configurable. We can do it with one
simple header line.

Also add a simple test for patch. Editing the expectation confirms that
my simple test works as expected.