Name Python package "rcl-lang"
I hesitated about "pyrcl" too, especially because that's what it's
called in the repository, but from a Python project perspective, if you
see "pyrcl" in the dependency list, if you don't already know what RCL
is, you don't know what that is. If you see "rcl-lang", that's a little
more descriptive. Also, it aligns with the rcl-lang.org domain.