Add script to run all fuzzers briefly
I am writing a release checklist, and on there is to run all fuzzers
briefly, something I forgot in the past. I run fuzz_source and
fuzz_smith all the time, but the others not that much. So I put it in
the playbook, and then the Bash onlineliner was turning into multiple
lines, so it's time to rewrite it in Python.

It doesn't provide the proper dictionary arguments, but it's good enough
for now.