Check "cargo doc --workspace" on CI
By making a flake check for it, so we can also check locally. This check
currently fails (because it was not enforced, so of course it broke).

And, what I learned the hard way, rebased here now:

We need to provide the full Rust sources, including generated
Tree-sitter files, and also we need to create $out when the check
passes. We can just put the docs there.