Correct list in doc comment
This was caught by Clippy on Rust 1.80. At first I thought it was a
false positive, because the line started with *, but for emphasis rather
than a list. But Clippy was right, Rustdoc merged the paragraph into the
list item, which was not intentional, and adding a blank line fixes it.