Cross-compile binaries with Nix
I want to add some prebuild binaries. Although you shouldn't trust
binaries from random GitHub repositories, it may help some people, and
also, if you want to do stuff like use RCL on CI, that's difficult when
it's not packaged. Installing a Rust toolchain is overkill then, and
also, building from source wastes time in the build, so if we can
provide a binary, that may be helpful.

Now that I'm using 'cargo-zigbuild' anyway to build the portable wheel,
we can apply the same technique to build portable binaries, and even to
cross-compile! The default glibc that Zig selects is not *as* old as the
2.17 from manylinux, but it's old enough to be widely compatible.