Use ':' in assertions rather than ','
Initially I used a comma after assertions, like Python. It was always
a bit weird (even in Python), it's a statement but it has arguments.
This made formatting akward too. I justified it for myself by saying
"it's just a function call, but without the parens". That's how we
currently format it.

Then yesterday I was skimming through the Jsonnet docs, and they use
a colon rather than comma, and that looks much more pleasant! I should
steal the good ideas; if RCL can fix list comprehension order w.r.t
Python, then it can also fix this one.

For compatibility, accept the comma as well (for now), but make the
formatter emit the colon. Similar to the colon after 'else', it does
not have to be a breaking change.