Format assertions like if, not like function call
This feels a bit more natural with the colon, to keep them on the same
line. We could go a step further and make the "assert cond" part its own
group, it can save 5 characters, but if you have a large expression like
that, we can also just break the expression, so I think this is okay.