Add many-to-many example
I've run into this situation many times, especially with user-group
memberships. It came up again at work today, which made me think a bit
more deeply about the topic. One possible way to address "to which axis
do you attach the relationship" is "to all of them", and then verify
that there are no inconsistencies. This is an exploration of what that
looks like in RCL.