Rename 'scalar' to 'single' in error messages
Scalar feels like too much jargon, and it's not really the right term
anyway, I just use it for lack of a better alternative, but we shouldn't
force it onto users.

'git grep scalar' confirms that the only remaining references to the
term are in the source code, or related to "Unicode scalar value", which
is an upstream term.