Point Helix grammar source at external repo
Helix needs the pre-compiled grammar, but if you have just a checkout of
the RCL repository, you don't have that. I never discovered this because
in my checkout of course I build the grammar. Fortunately Sergey
reported it so we can address this, thanks Sergey! Also fortunately, we
already have a repository with a compiled grammar, because Zed needs
this too, so we can point there.