Rephrase the changelog entries a bit for clarity
Unrelated to this change, but for the assertion syntax change, we should
mention that the formatter auto-updates documents.

Also, the list renders quite dense, an inline code span is not really
the place to demonstrate unpack. I'll leave that to a link. Maybe I
should do an announcement post as well with highlights, in addition to
the more mundane changelog. But RCL is doesn't have that many changes
going into a release anyway. For now I think it's okay.