Insert runtime typecheck for dict unpack if needed
In cases where the collection in '...collection' can't be statically
verified to be a dict, we need to do a runtime check. We can change the
expected type for the collection though.

Unfortunately this makes the error messages in case of a mismatch on
the type a bit more obscure. I think that at this point, that's an
acceptable cost to keep the implementation simpler.

This commit only handles dicts, but also lays the groundwork for
checking scalar unpack. For those, since it can be a list or set,
we need to move the typecheck to runtime, because there is no
single expected type to feed into the check. Well, we could feed
in Union[List[T], Set[T]], but the union is then conjured up by the
typechecker, which could make errors more difficult to understand, so
let's just do a runtime check.