Add --about that prints license info
For now it hard-codes the dependency list. I don't expect the list to
change often, so I think this is okay.

Possibly at some point we can use a tool like "cargo about" to generate
the list of depedencies aand licenses, but for a single dependency, I
think this is overkill.