Remove todo in scalar unpack error message
It is still kind of true, here is a reproducer:

    let xs = [1, 2]; { a = 1, ..xs }
    stdin:1:27
      ╷
    1 │ let xs = [1, 2]; { a = 1, ..xs }
      ╵                           ^~~~
    Error: Invalid unpack in dict.

    Help: '..' unpacks lists and sets, use '...' to unpack dicts.

I don't think the help is that bad though, the error does say "invalid
unpack" and the help explains that there are ".." and "...", the fact
that "xs" is a list and ".." is the right thing to unpack a list, well
yeah but what do you expect unpacking a list into a dict to do? So I
think it's fine, let's not complicate the already complex error
reporting logic any further.