Prevent hangs from deeply nested values in recursion
This is a hang discovered by the fuzzer. In the initial form, it
discovered it from a closure (where the environment capture made it
grow), but the same applies to lists. Putting those deeply nested values
in BTreeSets triggers some pathological worst-case comparison where
every comparison has to descend all the way.

This applies to all functions that can recursively build values. The
fuzzer discovered it for transitive_closure but in fact we can trigger
similar behavior with fold (which has never been found by the fuzzer!).
So guard that one as well.

You can still manually do recursion, but there is little we can do about
that, RCL is not really meant for it so serves you right if you trigger
it, though of course at some point the fuzzer is going to learn about
this and then we will have to mitigate it ...