Limit nesting depth more aggressively in fuzz mode
The fuzzer has discovered the concept of "exponential growth" and now
calls Set.transitive_closure with a function that at each iteration
creates a dict with *four* closures that each refer to the previous'
iteration's value. This feels like whack-a-mole, I can't make the depth
too small either. But let's see if maybe there is a sweet spot that is
acceptable.