Fix formatting of collections that contain unpacks
The formatter used to add a space at the end. That was an oversight when
I added the unpack feature. Fortunately it's an easy fix.

I added the golden test first to confirm it would have caught the bug.
Also, it turns out there was a different fmt golden tests that already
encoded the bad space. I missed it in review.