Delete Set.map and Set.flat_map
These functions turn out to be footguns, they bit me twice already in
the past weeks, and it's the kind of thing that is not obvious to debug.

Consider for example:

    let xs = {"apple", "banana", "orange"};
    let n = xs.map(x => x.len()).sum();

What do we expect n to be here? It looks like this is summing the length
of the strings (5, 6, and 6), but wrong! It sums all of the *unique*
lengths, so only 5 + 6.

Since this is quite the pitfall, I think it's better to delete the map
function. Better to be explicit. Functions to convert between list and
set would help to make the removal palatable.