Replace website tagline
One commenter on Lobste.rs pointed out that the "make X Y again" can
have political connotations. That was not the intent, and I don't want
to scare people away with that.

I brainstormed several options, in the end this one captures the same
spirit ("boring", it should not create problems or cause surprises, it
should not feel like solving puzzles, building things is exciting,
configuring them should be boring), and even makes the goal a bit
clearer by hinting at the remove-duplication part.